swell.py is refactored to commands.clone/create.py etc. The outcome did not speedup swell CLI, but simplified the source swell.py
