+++ src/swell/swell.py
@@ -1,32 +1,61 @@
-# (C) Copyright 2021- United States Government as represented by the Administrator of the
-# National Aeronautics and Space Administration. All Rights Reserved.
-#
-# This software is licensed under the terms of the Apache Licence Version 2.0
-# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
-
-
-# --------------------------------------------------------------------------------------------------
-
-
 import click
-from typing import Union, Optional, Literal
+import importlib
+import sys
 
-from swell.deployment.platforms.platforms import get_platforms
-from swell.deployment.create_experiment import clone_config, create_experiment_directory
-from swell.deployment.launch_experiment import launch_experiment
-from swell.tasks.base.task_base import task_wrapper, get_tasks
-from swell.test.test_driver import test_wrapper, valid_tests
-from swell.test.suite_tests.suite_tests import run_suite, TestSuite
-from swell.suites.all_suites import AllSuites
+# Import the package to access __version__
+import swell
+__version__ = swell.__version__
+
 from swell.utilities.welcome_message import write_welcome_message
-from swell.utilities.scripts.utility_driver import get_utilities, utility_wrapper
+
+COMMANDS = {
+    "clone":   "swell.commands.clone",
+    "create":  "swell.commands.create",
+    "launch":  "swell.commands.launch",
+    "t1test":  "swell.commands.t1test",
+    "t2test":  "swell.commands.t2test",
+    "task":    "swell.commands.task",
+    "test":    "swell.commands.test",
+    "utility": "swell.commands.utility",
+}
 
 
-# --------------------------------------------------------------------------------------------------
+class SwellCLI(click.MultiCommand):
+    def list_commands(self, ctx):
+        """Returns sorted list of command names for the Commands: section"""
+        return sorted(COMMANDS.keys())
+
+    def get_command(self, ctx, name):
+        """Lazy-load the requested command module and return its Click command object"""
+        if name not in COMMANDS:
+            raise click.UsageError(f"Unknown command: {name}")
+
+        module = importlib.import_module(COMMANDS[name])
+
+        # The command function is usually named the same as the subcommand
+        cmd_func_name = name.replace("-", "_")  # handles possible future hyphenated names
+        if hasattr(module, cmd_func_name):
+            return getattr(module, cmd_func_name)
+
+        # Fallback: look for any click.Command in the module
+        for attr_name in dir(module):
+            attr = getattr(module, attr_name)
+            if isinstance(attr, click.Command):
+                return attr
+
+        raise click.UsageError(
+            f"Command '{name}' not found in {COMMANDS[name]}. "
+            f"Expected a Click command named '{cmd_func_name}' or similar."
+        )
 
 
-@click.group()
-def swell_driver() -> None:
+@click.group(cls=SwellCLI)
+@click.version_option(
+    version=__version__,
+    prog_name="swell",
+    message="%(prog)s version %(version)s",
+)
+def swell_driver():
     """
     Welcome to swell!
 
@@ -40,266 +69,18 @@ def swell_driver() -> None:
     followed by
 
       swell launch <suite_path>
-
     """
     pass
 
 
-# --------------------------------------------------------------------------------------------------
-
-# Help strings for optional arguments
-
-input_method_help = 'Method by which to create the YAML configuration file. If choosing ' + \
-                    'defaults the setting for the default suite test will be used. If using ' + \
-                    'CLI you will be led through the questions to configure the experiment.'
-
-platform_help = 'If using defaults for input_method, this option is used to determine which ' + \
-                'platform to use for platform specific defaults. Options are ' + \
-                str(get_platforms())
-
-override_help = 'After generating the config file, parameters inside can be overridden ' + \
-                'using values from the override config file.'
-
-advanced_help = 'Show configuration questions which are otherwise not shown to the user.'
-
-no_detach_help = 'Tells the workflow manager not to detach. That is to say run the entire ' + \
-                 'run the entire workflow in the foreground and pass back a return code.'
-
-log_path_help = 'Directory to receive workflow manager logging output (instead of ' + \
-                '$HOME/cylc-run/<suite_name>)'
-
-datetime_help = 'Datetime to use for task execution. Format is yyyy-mm-ddThh:mm:ss. Note that ' + \
-                'non-numeric characters will be stripped from the string. Minutes and seconds ' + \
-                'are optional.'
-
-model_help = 'Data assimilation system. I.e. the model being initialized by data assimilation.'
-
-ensemble_help = 'When handling ensemble workflows using a parallel strategy, ' + \
-                'specify which packet of ensemble members to consider.'
-
-slurm_help = """
-Customize SLURM directives, globally (e.g., account name), for specific tasks,
-or for task-model combinations.
-"""
-
-skip_r2d2_help = """Skip registering this experiment and storing products in R2D2."""
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-@swell_driver.command()
-@click.argument('suite', type=click.Choice(AllSuites.config_names()))
-@click.option('-m', '--input_method', 'input_method', default='defaults',
-              type=click.Choice(['defaults', 'cli']), help=input_method_help)
-@click.option('-p', '--platform', 'platform', default='nccs_discover_sles15',
-              type=click.Choice(get_platforms()), help=platform_help)
-@click.option('-o', '--override', 'override', default=None, help=override_help)
-@click.option('-a', '--advanced', 'advanced', default=False, help=advanced_help)
-@click.option('-s', '--slurm', 'slurm', default=None, help=slurm_help)
-@click.option('-k', '--skip-r2d2', 'skip_r2d2', is_flag=True, default=False, help=skip_r2d2_help)
-def create(
-    suite: str,
-    input_method: str,
-    platform: str,
-    override: Union[dict, str, None],
-    advanced: bool,
-    slurm: str,
-    skip_r2d2: bool
-) -> None:
-    """
-    Create a new experiment
-
-    This command creates an experiment directory based on the provided suite name and options.
-
-    Arguments: \n
-        suite (str): Name of the suite you wish to run. \n
-
-    """
-
-    # Create the experiment directory
-    create_experiment_directory(suite, input_method, platform, override, advanced, slurm, skip_r2d2)
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-@swell_driver.command()
-@click.argument('configuration')
-@click.argument('experiment_id')
-@click.option('-m', '--input_method', 'input_method', default='defaults',
-              type=click.Choice(['defaults', 'cli']), help=input_method_help)
-@click.option('-p', '--platform', 'platform', default=None, help=platform_help)
-@click.option('-a', '--advanced', 'advanced', default=False, help=advanced_help)
-def clone(
-    configuration: str,
-    experiment_id: str,
-    input_method: str,
-    platform: str,
-    advanced: bool
-) -> None:
-    """
-    Clone an existing experiment
-
-    This command creates an experiment directory based on the provided experiment configuration.
-
-    Arguments: \n
-        configuration (str): Path to a YAML containing the experiment configuration you wish to
-        clone from. \n
-
-    """
-    # Create experiment configuration by cloning from existing experiment
-    experiment_dict_str = clone_config(configuration, experiment_id, input_method, platform,
-                                       advanced)
-
-    # Create the experiment directory
-    create_experiment_directory(experiment_dict_str)
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-@swell_driver.command()
-@click.argument('suite_path')
-@click.option('-b', '--no-detach', 'no_detach', is_flag=True, default=False, help=no_detach_help)
-@click.option('-l', '--log_path', 'log_path', default=None, help=log_path_help)
-def launch(
-    suite_path: str,
-    no_detach: bool,
-    log_path: str
-) -> None:
-    """
-    Launch an experiment with the cylc workflow manager
-
-    This command launches an experiment using the provided suite path and options.
-
-    Arguments: \n
-        suite_path (str): Path to where the flow.cylc and associated suite files are located. \n
-
-    """
-    launch_experiment(suite_path, no_detach, log_path)
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-@swell_driver.command()
-@click.argument('task', type=click.Choice(get_tasks()))
-@click.argument('config')
-@click.option('-d', '--datetime', 'datetime', default=None, help=datetime_help)
-@click.option('-m', '--model', 'model', default=None, help=model_help)
-@click.option('-p', '--ensemblePacket', 'ensemblePacket', default=None, help=ensemble_help)
-def task(
-    task: str,
-    config: str,
-    datetime: Optional[str],
-    model: Optional[str],
-    ensemblePacket: Optional[str]
-) -> None:
-    """
-    Run a workflow task
-
-    This command executes a task using the provided task name, configuration file and options.
-
-    Arguments:\n
-        task (str): Name of the task to execute.\n
-        config (str): Path to the configuration file for the task.\n
-
-    """
-    task_wrapper(task, config, datetime, model, ensemblePacket)
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-@swell_driver.command()
-@click.argument('utility', type=click.Choice(get_utilities()))
-def utility(utility: str) -> None:
-    """
-    Run a utility script
-
-    This command performs a utility operation specified by the utility argument.
-
-    Arguments:\n
-        utility (str): Name of the utility operation to perform.\n
-
-    """
-    utility_wrapper(utility)
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-@swell_driver.command()
-@click.argument('test', type=click.Choice(valid_tests))
-def test(test: str) -> None:
-    """
-    Run one of the test suites
-
-    This command performs the test specified by the test argument.
-
-    Arguments:\n
-        test (str): Name of the test to execute.
-
-    """
-    test_wrapper(test)
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-@swell_driver.command()
-@click.option('-p', '--platform', 'platform', type=click.Choice(get_platforms()),
-              default="nccs_discover_sles15", help=platform_help)
-@click.argument('suite', type=click.Choice(("hofx", "3dvar_marine", "3dvar_atmos",
-                                            "localensembleda", "3dvar_cycle")))
-def t1test(
-    suite: Literal["hofx", "3dvar_marine", "3dvar_atmos", "localensembleda", "3dvar_cycle"],
-    platform: Optional[str] = "nccs_discover_sles15"
-) -> None:
-    """
-    Run a particular swell suite from the tier 1 tests.
-
-    Arguments:
-        suite (str): Name of the suite to run (e.g., 3dvar_marine, 3dvar_atmos, localensembleda)
-    """
-    run_suite(suite, platform, TestSuite.TIER1)
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-@swell_driver.command()
-@click.option('-p', '--platform', 'platform', type=click.Choice(get_platforms()),
-              default="nccs_discover_sles15", help=platform_help)
-@click.argument('suite', type=click.Choice(("hofx", "3dvar_marine", "ufo_testing",
-                                            "convert_ncdiags", "3dfgat_atmos", "build_jedi")))
-def t2test(
-    suite: Literal["hofx", "3dvar_marine", "ufo_testing",
-                   "convert_ncdiags", "3dfgat_atmos", "build_jedi"],
-        platform: Optional[str] = "nccs_discover_sles15"
-) -> None:
-    """
-    Run a particular swell suite from the tier 2 tests.
-
-    Arguments:
-        suite (str): Name of the suite to run (e.g., hofx, 3dvar_marine, ufo_testing)
-    """
-    run_suite(suite, platform, TestSuite.TIER2)
-
-
-# --------------------------------------------------------------------------------------------------
-
-
 def main() -> None:
-    """
-    Main Function
+    if len(sys.argv) > 1:
+        first_arg = sys.argv[1]
+        if first_arg not in ("--help", "-h", "--version", "-V") and "--help" not in sys.argv:
+            write_welcome_message()
 
-    This function is the entry point for swell. It writes a welcome message and
-    sets up the driver group.
-    """
-    write_welcome_message()
     swell_driver()
 
 
-# --------------------------------------------------------------------------------------------------
+if __name__ == "__main__":
+    main()
